Fixes #51

+++ deemon/cmd/refresh.py
@@ -28,7 +28,7 @@ class Refresh:
         self.download_all = ignore_filters
         self.seen = None
 
-        if self.time_machine:
+        if self.time_machine and not self.waiting_for_refresh():
             logger.info(f":: Time Machine active: {datetime.strftime(self.time_machine, '%b %d, %Y')}!")
             config.set('by_release_date', False)
             self.db.remove_specific_releases({'tm_date': str(self.time_machine)})
